Separate options from cmds
